added  evaluation results in test.ipynb & save gt_counts in txt files
- added save_gt_counts & mv_val_split_to_train in density_map_generation for each dataset;
- added evaluation results in test.ipynb;
- added evaluation_results_as_df in evaluation.py;
- added predict_density_maps_and_get_counts in net_utils.py;
- added plot_gt_vs_pred_counts in vis.py;
- remark 1: gt_density_maps are generated in ./utils/preprocessing/misc.py with dtype=np.float64;
- remark 2: results with a high std among different trainings but with the same train/val/test split for vgg_cells (sensitivity to NN initialization? do we need BatchNorm?)

+++ utils/preprocessing/misc.py
@@ -3,7 +3,7 @@ import skimage
 
 
 def gaussian_smoothing(gt_dots, sigma):
-    gt_density_maps = np.zeros((len(gt_dots), *gt_dots[0].shape))
+    gt_density_maps = np.zeros((len(gt_dots), *gt_dots[0].shape), dtype=np.float64)
     
     for i in range(len(gt_dots)):
         gt_density_maps[i] = skimage.filters.gaussian(gt_dots[i],
